+++ GuateChances-API/src/main/java/ggnc/guatechancesapi/Models/Domain/Card.java
@@ -17,7 +17,7 @@ public class Card {
         this.csv = csv;
     }
 
-    public Card(int number) {
+    public Card(Long number) {
         this.number = number;
     }
